Disable Nagle on TCP socket: too slow for interactive
The wire-protocol is latency-bound, not bandwidth-constrained. So
disable the Nagle algorithm to prevent large delays from hurting our
performance too much.

NOTE: Considering that we are implementing a request-reply protocol it
might be possible to improve performance without disabling the Nagle
algorithm. E.g. by taking control over buffering ourselves instead of
leaving it up to the iostream class.