cucumber-cpp support for cgreen test framework.
As a user I want to use cgreen as an alternative test framework to gtest
and boost in order to test native C complete with mocking capability.

cucumber-cpp now recognizes and harnesses cgreen as a testing framework.
The cgreen reporting is intercepted and presented on the failure of a
cucumber
test.

An example (FizzBuzz) is written as an example of a testable C progam
and
is included in the examples directory.

Unit tests for the cgreen driver are modeled after the boost and gtest
drivers and are passing.