Add support for using ephemeral ports
Ephemeral ports are used when binding to port '0'. The operating system
then automatically selects a (non-zero) port for you that's unused at
the time. This is great to prevent conflicts.

This commit extends the wire server such that it reports the actual port
number we've bound to, even when using ephemeral ports. This is useful
to prevent port conflicts on CI systems running many tests in parallel.