Use dependencies via target_link_libraries exclusively
This allows all dependencies to be resolved in CMake's generation step
instead of before that time during script execution.

Additionally properties such as transitive dependencies are now handled
properly and include directories of dependencies automatically added to
the include path.

NOTE: This includes a workaround for cucumber/cucumber-cpp#190 which was
      originally discovered when doing this change.