Mark StepManager as a singleton
Because it already is. So make all its methods static to remove the need
for an instance to use it.

Additionally make the addStep() function return the ID of the step it
just registered and take its argument by value instead of by reference.
The former change improves the usability of this function: you don't
need to keep a pointer around just to be able to get the ID later on.
The latter allows the optimizer to elide a copy by taking ownership
instead of a non-owning reference.