Support building as a shared library
On Windows, it is necessary to add the __declspec() attribute around
some functions/classes in order to correctly link cucumber-cpp as a
shared library. This commit adds the makes it possible for users to
easily build cucumber-cpp either as a static or dynamic library by
defining either CUKE_ENABLE_SHARED_LIB when configuring with CMake.