Add a step-matching hook
The motivation behind this commit is that we get too many ambiguous
step matchings because we use (.+) as view expressions. What's a view
and what isn't is only determined at runtime (through the view
evaluators which register at runtime), so we can't describe it in a
regular expression.

With this feature, we have a say in whether or not a step should
match, by providing our own matching hook. We use regex's named
capture mechanism to activate our own matching algorithm.

Credit to @stk-ableton for the original patch.