Support unpacking regex captures to step-function arguments
On C++11 (and beyond), support extracting of the regex' captures to the
arguments of the step function. This enables a more natural syntax to
receive these than using the REGEX_PARAM macro.

Currently all captures from the regex that aren't passed as parameters
can still be extracted by means of the REGEX_PARAM macro.

In the future it may be desirable to check, at compile-time, that the
amount of captures in the regex equals the amount of function
parameters. Unfortunately this is relatively doable in C++17, because of
the presence of constexpr lambdas, more difficult in C++14, and probably
impossible in C++11.