Simplify decoding of function signature
This gets rid of the FunctionSignature and FunctionArgs helper classes
by using a pointer-to-member-function and being able to get the
arguments as a pack directly. As a result the function signature no
longer has to be specified explicitly as a template parameter, it's
instead determined by the compiler using template type deduction.

This further simplifies the C++98 code to be ridiculously
straightforward and the C++11 is simpler because it no longer has to be
C++98 compatible.