Don't use shared_ptr for tag expressions
Also make it const-correct and use a vector for storing subexpressions.

We don't need shared_ptr at all because we always have unique and direct
ownership and never have polymorphism.

We can use a vector instead of linked list because there's no reason to
go with a linked list. Vector has better memory locality, better
allocation performance (when the size is known ahead of time, which it
is) and is simpler to use.