Build with warnings, ignoring json_spirit
Move json_spirit to a separate directory, out of the default
include-path. This so that we can mark that include-path as being a
"system include". This prevents at least GCC and Clang from producing
warnings in these files.

Then enable the -Wall and -Wextra warning sets by default for both GCC
and Clang.