MSVC: Don't warn about use of "unsafe" standard functions
Mostly this is bullshit and this warning only adds noise because it's a
false-positive most of the time. Instead it promotes use of non-portable
Microsoft-only functions.

Yes in theory they've been standardized in C11, but we use C++, and none
of the C++ standards has adopted this part of the C standard yet.
Furthermore it's an _optional_ part of the standard (Annex K), which
doesn't help with portability due to its optional nature.