Link to (position independent) dependencies when building a shared lib
Making sure GTest/GMock is built as position independent code is
necessary when it's linked statically and we try to link it into
Cucumber-CPP when that's built as a shared library.

Furthermore Boost::unit_test_framework is a direct dependency of
cucumber-cpp: we need to link to it when we use it.