Hide all symbols by default
This prevents unmarked symbols, which we consider to be internals, from
being available on the dynamic library's dynamic symbol table (EXPORT
table on Windows).

Furthermore we split out a static library. This allows us to expose the
library internals to unit tests only, through this static library. This
works because symbol hiding only affects dynamic linking, not static
linking.