Fix build: don't propagate PRIVATEly used INTERFACE libs
Because, for STATIC libraries, CMake still does (some) propagation of
PRIVATE dependencies installing a STATIC library linking PRIVATEly to a
library requires that library to be installed too. This however, should
not be necessary for INTERFACE libraries if and only if they don't have
an affect on the link time interface. Given that CMake, currently, isn't
smart enough to figure this out by itself: just manually extract the
properties we want from it.

This fixes #193