only enable minimal set of options by default
This is to reduce error output since cmake fails when an option is enabled
but a dependency for it not installed.

Test frameworks are off by default because every one has an external
dependency. A warning is shown when no test framework is selected as one is
needed for cucumber-cpp to be usable.